requirements_pip.txt
ansi2html==1.9.1
asttokens==3.0.0
attrs==24.3.0
blas==1.0
blinker==1.9.0
brotli-python==1.0.9
bzip2==1.0.8
ca-certificates==2025.2.25
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
dash==2.14.2
debugpy==1.8.11
decorator==5.1.1
executing==0.8.3
expat==2.6.4
flask==3.0.3
flask-compress==1.13
icc_rt==2022.1.0
idna==3.7
importlib-metadata==8.5.0
iniconfig==1.1.1
intel-openmp==2023.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
libffi==3.4.4
libsodium==1.0.18
markupsafe==3.0.2
matplotlib-inline==0.1.6
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.0.1
numpy-base==2.0.1
openssl==3.0.16
packaging==24.2
parso==0.8.4
pip==25.0
platformdirs==4.3.7
plotly==5.24.1
pluggy==1.5.0
prompt-toolkit==3.0.43
prompt_toolkit==3.0.43
psutil==5.9.0
pure_eval==0.2.2
pygments==2.19.1
pyserial==3.5
pysocks==1.7.1
pytest==8.3.4
python==3.12.9
python-dateutil==2.9.0post0
python-fastjsonschema==2.20.0
pywin32==308
pyzmq==26.2.0
referencing==0.30.2
requests==2.32.3
retrying==1.3.3
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sqlite==3.45.3
stack_data==0.2.0
tbb==2021.8.0
tenacity==9.0.0
threadpoolctl==3.5.0
tk==8.6.14
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025a
urllib3==2.3.0
vc==14.42
vs2015_runtime==14.42.34433
wcwidth==0.2.5
werkzeug==3.0.6
wheel==0.45.1
win_inet_pton==1.1.0
xz==5.6.4
zeromq==4.3.5
zipp==3.21.0
zlib==1.2.13